+++ game/Headquarters.cpp
@@ -352,7 +352,7 @@ void StructureBuildForm::SetOwner(BuilderGob *pgobOwner)
 
 			int nStripIcon = puntc->panid->GetStripIndex("icon");
 			if (nStripIcon != -1) {
-				plstc->Add(puntc->panid, nStripIcon, 0, (void *)(int)puntc->ut, fDisabled);
+				plstc->Add(puntc->panid, nStripIcon, 0, (void *)(uintptr_t)puntc->ut, fDisabled);
 			}
 		}
 	}
